Move fixtures to fixtures.py
